+++ muse2/muse/audioframe_t.h
@@ -1,8 +1,10 @@
 #ifndef __AUDIOFRAME_T_H__
 #define __AUDIOFRAME_T_H__
 
+#include <stdint.h>
+
 namespace MusECore {
-	typedef unsigned audioframe_t;
+	typedef uint64_t audioframe_t;
 }
